Remove pandoc-commonmark.md proposal
Because there is a strong need for this (not just in pandoc but in Hackage and Haddock), I have been writing a native Haskell commonmark parser, and it is pretty close to being finished.  So this is no longer a promising SoC idea. If students are interested in working on pandoc, though, I'd be happy to discuss ideas with them.