Update ghc-code-generator.md
I'd like to be a resource for this too.

I'm primarily interested in making sure any work that touches user facing vectorization/simd api  or rounding semantics is done nicely. I've a passing familiarity with a lot of the other tasks that are proposed, but in terms of general compiler/code gen concerns Ben, Moritz and Matthew Pickering are great!