Add quickcheck-state-machine proposal.
Based on the experience from last year it seemed students had problems
finding a good library or application to test, so I've added that
choosing a toy library or application from a commonly used Haskell
resource (e.g. a tutorial, book, or blog post) is also fine.

I also added an alternative intermediate path where the student could
try to turn the state machine model into a mock and use it to test
software that depends of the original software.

Feedback most welcome!